Add native bridge build support and enhance publishing scripts
- Introduced a new `build-native.ps1` script to facilitate the building of native bridge assets for Windows and Linux platforms, including `mcu_serial_bridge.dll` and `libmcu_serial_bridge.so`.
- Updated `CoralinkerHost.csproj` and `CoralinkerSDK.csproj` to reference the new native assets and ensure proper inclusion in the build process.
- Enhanced `publish-host.ps1` to include a `-SkipNativeBuild` option, allowing users to bypass native asset compilation if already built.
- Updated `README.md` to document the new native bridge build process and usage instructions for the publishing scripts.
- Added `.gitattributes` to manage line endings for shell scripts in the packaging directory.

+++ MCUSerialBridge/c_core/src/msb_thread.c
@@ -4,7 +4,6 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
-#include <windows.h>
 
 #include "c_core_common.h"
 #include "msb_handle.h"
@@ -548,8 +547,8 @@ DWORD WINAPI parse_thread_func(LPVOID param)
                                 handle->fatal_error_callback_ctx);
                     }
                 } else {
-                    DBG_PRINT("Fatal Error: Duplicate within 5s (elapsed=%llums), seq=%u skipped", 
-                              elapsed_ms, seq);
+                    DBG_PRINT("Fatal Error: Duplicate within 5s (elapsed=%llums), seq=%u skipped",
+                              (unsigned long long)elapsed_ms, seq);
                 }
             } else {
                 // -------------------------------
